Debug
1. parser.js
- parseFloat for <alter> in both parse functions (was parseInt — 0.5 → 0 silently)

2. renderer.js
- Octave -2 now draws two dots below the number (was completely ignored)

3. app.js
- MIDI chord deduplication — simultaneous notes are now grouped by tick and only the highest pitch (melody note) is kept
4. app.js
- Key/time signature in renderSelectedPart now reads from the selected part's sub-doc with a fallback to the global doc

5. app.js
- parseFloat for <alter> in the melody auto-detection scoring loop

+++ js/parser.js
@@ -59,7 +59,7 @@ function parseXMLToJianpu(xmlDoc) {
                     const step = pitchNode.getElementsByTagName("step")[0].textContent;
                     let alter = 0;
                     const alterNode = pitchNode.getElementsByTagName("alter")[0];
-                    if (alterNode) alter = parseInt(alterNode.textContent);
+                    if (alterNode) alter = parseFloat(alterNode.textContent);
                     const octave = parseInt(pitchNode.getElementsByTagName("octave")[0].textContent);
 
                     let noteSemi = pitchToSemitones(step, alter, octave);
@@ -253,7 +253,7 @@ function parseXMLToNoteObjects(xmlDoc) {
                     const step = pitchNode.getElementsByTagName("step")[0].textContent;
                     let alter = 0;
                     const alterNode = pitchNode.getElementsByTagName("alter")[0];
-                    if (alterNode) alter = parseInt(alterNode.textContent);
+                    if (alterNode) alter = parseFloat(alterNode.textContent);
                     const octave = parseInt(pitchNode.getElementsByTagName("octave")[0].textContent);
 
                     let noteSemi = pitchToSemitones(step, alter, octave);
